requirements.txt
langchain
streamlit
openai
psycopg2-binary
sqlalchemy
langchain-community
langchain-experimental
langgraph
python-dotenv
langchain-openai
langchain_tavily
bs4
langgraph_swarm
fastapi
uvicorn
pydantic
flask
langchain_postgres
gunicorn
flask-cors